backup/requirements.txt
requests==2.28.1
pytz==2022.1
PyYAML==6.0